restructured guis
Made a common mainwindow where users can show a persistent frame converter window and/or a persistent stack converter window. The Log window appears in the main selection window rather than in the separate windows.
