+++ mib2hspy/gui/__init__.py
@@ -1,3 +1,5 @@
-from .gui import *
+from .stack_converter import *
 from .guiTools import *
-from .converter_gui import run_converter_gui
+from .frame_converter import run_converter_gui
+from .stack_converter import run_gui
+from .gui import gui
