+++ mib2hspy/gui/source/mib2hspy/main.cpp
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "stack_converter.h"
 
 #include <QApplication>
