+++ mib2hspy/gui/source/mib2hspy/stack_converter.cpp
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "stack_converter.h"
 #include "ui_mainwindow.h"
 
 MainWindow::MainWindow(QWidget *parent)
